requirements.txt
streamlit
twitchio
python-dotenv
asyncio